renamed collector_list module to collector_list_screen
