+++ lib/src/features/collectors/screen/collector_list_screen.dart
